refactor(createContentClient): moved function into own file
