+++ src/lib/ContentClient.spec.ts
@@ -33,48 +33,4 @@ describe('ContentClient', () => {
       );
     });
   });
-
-  context('createContentClient', () => {
-    it('should default the base url', () => {
-      const client = (<any>new ContentClient({
-        account: 'test'
-      })).contentClient;
-
-      expect(client.defaults.baseURL).to.eq('https://c1.adis.ws');
-    });
-
-    it('should use the override base url if provided', () => {
-      const client = (<any>new ContentClient({
-        account: 'test',
-        baseUrl: 'http://localhost:3000'
-      })).contentClient;
-      expect(client.defaults.baseURL).to.eq('http://localhost:3000');
-    });
-
-    it('should use the staging environment url if provided', () => {
-      const client = (<any>new ContentClient({
-        account: 'test',
-        baseUrl: 'http://localhost:3000',
-        stagingEnvironment:
-          'fhboh562c3tx1844c2ycknz96-gvzrfgnzc-1546264721816.staging.bigcontent.io'
-      })).contentClient;
-
-      expect(client.defaults.baseURL).to.eq(
-        'https://fhboh562c3tx1844c2ycknz96-gvzrfgnzc-1546264721816.staging.bigcontent.io'
-      );
-    });
-
-    it('should use the provided adaptor to send requests', () => {
-      const client = (<any>new ContentClient({
-        account: 'test',
-        adaptor: () =>
-          Promise.resolve<any>({
-            data: 'response'
-          })
-      })).contentClient;
-      return client.get('/test').then(response => {
-        expect(response).to.deep.eq({ data: 'response' });
-      });
-    });
-  });
 });
